weather-api/pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = "A weather API that fetches and returns weather data from a 3rd party API."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-limiter[redis]>=4.0.0",
    "python-dotenv>=1.2.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "responses>=0.25.8",
]
